requirements.txt
beautifulsoup4==4.10.0
matplotlib==3.4.3
matplotlib_label_lines==0.4.1
MiniSom==2.2.9
networkx==2.6.3
numpy==1.21.3
pyodbc==4.0.32
python-dotenv==0.19.2
python_dateutil==2.8.2
pyvis==0.1.9
requests==2.22.0
scikit_learn==1.0.1
scipy==1.7.1
snscrape==0.3.5.dev138+ga6b6f3f
tslearn==0.5.2
tweepy==4.1.0
